+++ src/Spreads.Core/Serialization/IDirectBuffer.cs
@@ -234,7 +234,7 @@ namespace Spreads.Serialization
         unsafe long InterlockedCompareExchangeInt64(int index, long value, long comparand);
         unsafe int VolatileReadInt32(int index);
         unsafe void VolatileWriteInt32(int index, int value);
-        unsafe int VolatileReadInt64(int index);
-        unsafe void VolatileWriteInt64(int index, int value);
+        unsafe long VolatileReadInt64(int index);
+        unsafe void VolatileWriteInt64(int index, long value);
     }
 }
